Rename LegalTerms.tsx to LegalTermsModal.tsx
