+++ apps/passport-client/components/modals/LegalTermsModal.tsx
