+++ apps/passport-client/components/modals/Modal.tsx
@@ -11,7 +11,7 @@ import { ChangedPasswordModal } from "./ChangedPasswordModal";
 import { ConfirmSkipSetupModal } from "./ConfirmSkipSetupModal";
 import { InfoModal } from "./InfoModal";
 import { InvalidUserModal } from "./InvalidUserModal";
-import { LegalTermsModal } from "./LegalTerms";
+import { LegalTermsModal } from "./LegalTermsModal";
 import { RequireAddPasswordModal } from "./RequireAddPasswordModal";
 import { ResolveSubscriptionErrorModal } from "./ResolveSubscriptionError";
 import { SettingsModal } from "./SettingsModal";
